move ResourcePool to types module
